refactor(server): llm cache to llm factory
