+++ apps/server/api/tests/llm_test.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-    use api::ws::llm::llm_cache::LlmCache;
+    use api::ws::llm::LlmFactory;
     use tokio_stream::StreamExt;
     use tracing::{error, info};
     use tracing_test::traced_test;
@@ -10,9 +10,8 @@ mod tests {
     #[ignore]
     /// cargo test --test llm_test --features cuda -- tests::test_llm_chat --ignored --show-output
     async fn test_llm_chat() {
-        LlmCache::init().await;
-        let llm = LlmCache::global().instance.clone();
-        let llm = llm.lock().await;
+        LlmFactory::init().await;
+        let llm = LlmFactory::global().get_llm();
         let mut output = llm.chat("你是一个助手，所有回答必须使用纯文本自然语言，禁止使用任何Markdown符号如#、-、*等并且数字使用中文字代替。".to_string()
             ,String::from("静夜思的内容"));
         let mut result = Vec::new();
@@ -36,9 +35,8 @@ mod tests {
     #[ignore]
     /// cargo test --test llm_test --features cuda -- tests::test_llm_short_question --ignored --show-output
     async fn test_llm_short_question() {
-        LlmCache::init().await;
-        let llm = LlmCache::global().instance.clone();
-        let llm = llm.lock().await;
+        LlmFactory::init().await;
+        let llm = LlmFactory::global().get_llm();
         let mut output = llm.chat("你是一个助手，所有回答必须使用纯文本自然语言，禁止使用任何Markdown符号如#、-、*等并且数字使用中文字代替。".to_string()
             ,String::from("1+1="));
         let mut result = Vec::new();
